requirements.txt
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow
python-dotenv
psycopg2
flake8
requests
celery
redis
googletrans==4.0.0-rc1
drf_social_oauth2==2.2.0
